requirements.txt
django==3.2.0
django-model-utils==4.2.0
djangorestframework==3.12.4
celery[redis]==5.3.1
psycopg2-binary==2.9